Moved fake grafana client
